Create new files and split integration tests
