+++ test/integration/grpc_management_plane_test_helpers.go
@@ -1,0 +1,85 @@
+// Copyright (c) F5, Inc.
+//
+// This source code is licensed under the Apache License, Version 2.0 license found in the
+// LICENSE file in the root directory of this source tree.
+
+package integration
+
+import (
+	"fmt"
+	"net/http"
+	"testing"
+	"time"
+
+	"github.com/go-resty/resty/v2"
+	"github.com/stretchr/testify/assert"
+	"github.com/stretchr/testify/require"
+)
+
+const (
+	retryCount       = 5
+	retryWaitTime    = 4 * time.Second
+	retryMaxWaitTime = 5 * time.Second
+)
+
+var mockManagementPlaneAPIAddress string
+
+func performConfigApply(t *testing.T, nginxInstanceID string) {
+	t.Helper()
+
+	client := resty.New()
+	client.SetRetryCount(retryCount).SetRetryWaitTime(retryWaitTime).SetRetryMaxWaitTime(retryMaxWaitTime)
+
+	url := fmt.Sprintf("http://%s/api/v1/instance/%s/config/apply", mockManagementPlaneAPIAddress, nginxInstanceID)
+	resp, err := client.R().EnableTrace().Post(url)
+
+	require.NoError(t, err)
+	assert.Equal(t, http.StatusOK, resp.StatusCode())
+}
+
+func performInvalidConfigApply(t *testing.T, nginxInstanceID string) {
+	t.Helper()
+
+	client := resty.New()
+	client.SetRetryCount(retryCount).SetRetryWaitTime(retryWaitTime).SetRetryMaxWaitTime(retryMaxWaitTime)
+
+	body := fmt.Sprintf(`{
+			"message_meta": {
+				"message_id": "e2254df9-8edd-4900-91ce-88782473bcb9",
+				"correlation_id": "9673f3b4-bf33-4d98-ade1-ded9266f6818",
+				"timestamp": "2023-01-15T01:30:15.01Z"
+			},
+			"config_apply_request": {
+				"overview": {
+					"files": [{
+						"file_meta": {
+							"name": "/etc/nginx/nginx.conf",
+							"hash": "ea57e443-e968-3a50-b842-f37112acde71",
+							"modifiedTime": "2023-01-15T01:30:15.01Z",
+							"permissions": "0644",
+							"size": 0
+						},
+						"action": "FILE_ACTION_UPDATE"
+					},
+					{
+						"file_meta": {
+							"name": "/unknown/nginx.conf",
+							"hash": "bd1f337d-6874-35ea-9d4d-2b543efd42cf",
+							"modifiedTime": "2023-01-15T01:30:15.01Z",
+							"permissions": "0644",
+							"size": 0
+						},
+						"action": "FILE_ACTION_ADD"
+					}],
+					"config_version": {
+						"instance_id": "%s",
+						"version": "6f343257-55e3-309e-a2eb-bb13af5f80f4"
+					}
+				}
+			}
+		}`, nginxInstanceID)
+	url := fmt.Sprintf("http://%s/api/v1/requests", mockManagementPlaneAPIAddress)
+	resp, err := client.R().EnableTrace().SetBody(body).Post(url)
+	require.NoError(t, err)
+	assert.Equal(t, http.StatusOK, resp.StatusCode())
+}
